+++ jobs/v4/howto/list_job_sample.go
@@ -1,0 +1,62 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package howto
+
+// [START job_search_list_job]
+import (
+	"context"
+	"fmt"
+	"io"
+
+	talent "cloud.google.com/go/talent/apiv4beta1"
+	"google.golang.org/api/iterator"
+	talentpb "google.golang.org/genproto/googleapis/cloud/talent/v4beta1"
+)
+
+// listJobs lists jobs with a filter, for example
+// `companyName="projects/my-project/companies/123"`.
+func listJobs(w io.Writer, projectID, filter string) error {
+	ctx := context.Background()
+
+	// Initialize a jobService client.
+	c, err := talent.NewJobClient(ctx)
+	if err != nil {
+		fmt.Printf("talent.NewJobClient: %v\n", err)
+		return err
+	}
+
+	// Construct a listJobs request.
+	req := &talentpb.ListJobsRequest{
+		Parent: "projects/" + projectID,
+		Filter: filter,
+	}
+
+	it := c.ListJobs(ctx, req)
+
+	for {
+		resp, err := it.Next()
+		if err == iterator.Done {
+			return nil
+		}
+		if err != nil {
+			fmt.Printf("it.Next: %v\n", err)
+			return err
+		}
+		fmt.Fprintf(w, "Listing job: %v\n", resp.GetName())
+		fmt.Printf("Job title: %v\n", resp.GetTitle())
+	}
+}
+
+// [END job_search_list_jobs]
